backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
requests==2.31.0
pandas==1.5.3
yfinance==0.2.36
newsapi-python==0.2.7
python-jose==3.3.0
pydantic==2.6.1
numpy==1.24.3 